Add X-Api-Object-Id request header
This header is the inverse of the response header of the same name.
Whilst the response header returns the Id of the last object returned,
the request header is used to control the next object that is
returned. Its main utility is in testing - by setting this header
to -1 you can ensure that you always get back the entire queryset.