Add zap id to the subscription
The original plan would have meant that all of a user's subscriptions
(i.e. zaps) would have been active / inactive at the same time, which is
not what you want. A user can create two separate zap that both use the
same trigger, and these need to be independent. The updated version uses
the zap id to distinguish between zaps.

As a side note, the zap id changes each time you subscribe / unsubscribe
which is not what I had expected. As a result I have simplified some of
the code as there is no "resubscribe" concept - each time a zap is
turned back on it's a new zap, and a new subscription.